# Optimize string transformation using `String#tr`