@tagliala 2024-08-24
### Before (main branch): 
```
$ ./bin/bench bench benchmark/*_attribute.haml 
=================================================
 Rendering: benchmark/boolean_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0   118.170k i/100ms
-------------------------------------------------
         haml v6.3.0      1.723M i/s (0.001ms) -      8.626M
=================================================
 Rendering: benchmark/class_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    41.452k i/100ms
-------------------------------------------------
         haml v6.3.0    472.804k i/s (0.002ms) -      2.404M
=================================================
 Rendering: benchmark/common_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    88.967k i/100ms
-------------------------------------------------
         haml v6.3.0      1.175M i/s (0.001ms) -      5.961M
=================================================
 Rendering: benchmark/data_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0     8.910k i/100ms
-------------------------------------------------
         haml v6.3.0     91.181k i/s (0.011ms) -    463.320k
=================================================
 Rendering: benchmark/dynamic_boolean_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0   139.402k i/100ms
-------------------------------------------------
         haml v6.3.0      2.230M i/s (0.000ms) -     11.152M
=================================================
 Rendering: benchmark/id_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    43.752k i/100ms
-------------------------------------------------
         haml v6.3.0    483.908k i/s (0.002ms) -      2.450M
```

### After (this branch):
```
$ ./bin/bench bench benchmark/*_attribute.haml 
=================================================
 Rendering: benchmark/boolean_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0   122.914k i/100ms
-------------------------------------------------
         haml v6.3.0      1.778M i/s (0.001ms) -      8.973M
=================================================
 Rendering: benchmark/class_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    41.359k i/100ms
-------------------------------------------------
         haml v6.3.0    478.132k i/s (0.002ms) -      2.399M
=================================================
 Rendering: benchmark/common_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    90.165k i/100ms
-------------------------------------------------
         haml v6.3.0      1.195M i/s (0.001ms) -      6.041M
=================================================
 Rendering: benchmark/data_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    10.637k i/100ms
-------------------------------------------------
         haml v6.3.0    110.123k i/s (0.009ms) -    553.124k
=================================================
 Rendering: benchmark/dynamic_boolean_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0   145.471k i/100ms
-------------------------------------------------
         haml v6.3.0      2.287M i/s (0.000ms) -     11.492M
=================================================
 Rendering: benchmark/id_attribute.haml
=================================================
Calculating -------------------------------------
         haml v6.3.0    42.905k i/100ms
-------------------------------------------------
         haml v6.3.0    490.843k i/s (0.002ms) -      2.488M
```